third_party/blink/web_tests/http/tests/security/window-onerror-exception-in-iframe.html

!<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>

<script>
  errors = [];
  iframeLoaded = false;
  window.addEventListener('error',(e) => {
    errors.push(e);
    e.preventDefault();
  });
  window.addEventListener('message', (e) => {
    if (e.data === 'IFrameLoaded') {
      iframeLoaded = true;
    } else {
      errors.push('Got errant message');
    }
  });
</script>

<iframe src="http://localhost:8000/security/resources/onerror-iframe.html"></iframe>

<script>
promise_test(async (t) => {
  if (!iframeLoaded) {
    await new Promise(resolve => window.addEventListener('message',resolve));
  }
  assert_true(iframeLoaded);
  assert_array_equals(errors,[],'No errors should get reported to window');
},'Test that exceptions in iframe are not reported to the main frame window.onerror handler');
</script>
